+++ CHANGELOG.md
@@ -1,6 +1,46 @@
 Changelog
 ===
 
+## Version 11.11.0
+### ✨ New & Improved
+
+* correctly parse malformed md syntax ([#1279](https://github.com/readmeio/markdown/issues/1279)) ([4bbd8fb](https://github.com/readmeio/markdown/commit/4bbd8fbb70ac1a5e96ebec2866957f34a88a33d0))
+
+### 🛠 Fixes & Updates
+
+* **mdxish:** can't render RC with snake_case name ([#1281](https://github.com/readmeio/markdown/issues/1281)) ([0e50d1c](https://github.com/readmeio/markdown/commit/0e50d1c03c0e964273b34f18ad8660421bc12495))
+* **extractMagicBlocks:** ensure magic blocks are surrounded by newlines ([#1286](https://github.com/readmeio/markdown/issues/1286)) ([67a6914](https://github.com/readmeio/markdown/commit/67a6914ae3207584cee265c0b850977e9b6b95e9))
+* **mdxish:** fix rendering callout from magic blocks ([#1275](https://github.com/readmeio/markdown/issues/1275)) ([171a5d8](https://github.com/readmeio/markdown/commit/171a5d89cc98ff25461e3f92a3c940895bd05bdc))
+* **mdxish:** jsx braces not evaluating ([#1256](https://github.com/readmeio/markdown/issues/1256)) ([f9fe3b7](https://github.com/readmeio/markdown/commit/f9fe3b7bc7c467b923b6f00e2553f88569a5200a))
+
+## Version 11.10.1
+### 🛠 Fixes & Updates
+
+* cardssssssss ([#1278](https://github.com/readmeio/markdown/issues/1278)) ([2066a8e](https://github.com/readmeio/markdown/commit/2066a8e2652636ee03b5d7449ca0a13441c65910)), closes [#1254](https://github.com/readmeio/markdown/issues/1254)
+
+## Version 11.10.0
+### ✨ New & Improved
+
+* **mdxish:** support custom component tag extraction with mdxish syntax ([#1277](https://github.com/readmeio/markdown/issues/1277)) ([2974ccf](https://github.com/readmeio/markdown/commit/2974ccfb064da520cfad8dc93088a5e2cb9df31b))
+
+## Version 11.9.4
+### 🛠 Fixes & Updates
+
+* **sanitizeComments:** handle magic block on the first line of document ([#1276](https://github.com/readmeio/markdown/issues/1276)) ([0a5ea94](https://github.com/readmeio/markdown/commit/0a5ea94c160524f510e74b01fa0804b6fbd92056))
+
+## Version 11.9.3
+### 🛠 Fixes & Updates
+
+* card styles ([#1271](https://github.com/readmeio/markdown/issues/1271)) ([700f9ef](https://github.com/readmeio/markdown/commit/700f9ef1e716c1497b9f12a384b9cb793334afc6))
+
+## Version 11.9.2
+### 🛠 Fixes & Updates
+
+* add release instructions ([#1261](https://github.com/readmeio/markdown/issues/1261)) ([92dc651](https://github.com/readmeio/markdown/commit/92dc651ca8a6a61c6b73b77197df7bd62cbcaea7))
+* **mdxish:** parse table cells in magic blocks  ([#1255](https://github.com/readmeio/markdown/issues/1255)) ([0963562](https://github.com/readmeio/markdown/commit/0963562edac3676605cc517fc4512190a19cbb0c))
+* unblock releases update docs ([#1268](https://github.com/readmeio/markdown/issues/1268)) ([e7e68fa](https://github.com/readmeio/markdown/commit/e7e68fa9d4830d39c8eee356159416f712c54b5c))
+* update release for rdme v10 ([#1266](https://github.com/readmeio/markdown/issues/1266)) ([5859571](https://github.com/readmeio/markdown/commit/5859571cbbd80c77237f1ba0b97b9bc671a88439)), closes [#migrating-to-rdme9](https://github.com/readmeio/markdown/issues/migrating-to-rdme9)
+
 ## Version 11.9.1
 ### 🛠 Fixes & Updates